cp: do protocol.MustRegisterComponent when comp init instead of base.DefaultProvider.Init
