+++ providers/component-protocol/cpregister/base/provider_name_test.go
