--- providers/component-protocol/examples/components/parallel_demo/grid/provider.go
@@ -1,57 +1,0 @@
-// Copyright (c) 2021 Terminus, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package grid
-
-import (
-	"context"
-
-	"github.com/erda-project/erda-infra/base/servicehub"
-	"github.com/erda-project/erda-infra/providers/component-protocol/cptype"
-	"github.com/erda-project/erda-infra/providers/component-protocol/protocol"
-)
-
-type provider struct{}
-
-func (p *provider) Render(ctx context.Context, c *cptype.Component, scenario cptype.Scenario, event cptype.ComponentEvent, gs *cptype.GlobalStateData) error {
-	c.Data = map[string]interface{}{
-		"aa": "bb",
-	}
-	return nil
-}
-
-// Init .
-func (p *provider) Init(ctx servicehub.Context) error {
-	compName := "grid"
-	if ctx.Label() != "" {
-		compName = ctx.Label()
-	}
-	protocol.MustRegisterComponent(&protocol.CompRenderSpec{
-		Scenario: "parallel-demo",
-		CompName: compName,
-		RenderC:  func() protocol.CompRender { return &provider{} },
-	})
-	return nil
-}
-
-// Provide .
-func (p *provider) Provide(ctx servicehub.DependencyContext, args ...interface{}) interface{} {
-	return p
-}
-
-func init() {
-	servicehub.Register("component-protocol.components.parallel-demo.grid", &servicehub.Spec{
-		Creator: func() servicehub.Provider { return &provider{} },
-	})
-}
